code style change
